backend/requirements/base.txt
Django==3.2.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
drf-friendly-errors==0.14
drf-yasg==1.20.0
drf-yasg[validation]
psycopg2-binary==2.8.6
python-dotenv==0.17.1
requests==2.25.1
